+++ service/src/services/communication-adapter/communication.adapter.ts
@@ -1,16 +1,11 @@
-import { ConfigurationTypes, LogContext } from '@common/enums';
-import {
-  MatrixEntityNotFoundException,
-  NotEnabledException,
-} from '@common/exceptions';
+import { LogContext } from '@common/enums';
+import { MatrixEntityNotFoundException } from '@common/exceptions';
 import { Inject, Injectable, LoggerService } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
 import { MatrixAgentPool } from '@services/matrix/agent-pool/matrix.agent.pool';
 import { HistoryVisibility, JoinRule, MatrixClient } from 'matrix-js-sdk';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { MatrixRoomAdapter } from '@services/matrix/adapter-room/matrix.room.adapter';
 import { MatrixUserAdapter } from '@services/matrix/adapter-user/matrix.user.adapter';
-import { IOperationalMatrixUser } from '@services/matrix/adapter-user/matrix.user.interface';
 import { MatrixAgent } from '@services/matrix/agent/matrix.agent';
 import { MatrixAgentService } from '@services/matrix/agent/matrix.agent.service';
 import { MatrixUserManagementService } from '@services/matrix/management/matrix.user.management.service';
@@ -30,65 +25,36 @@ import { RoomDeleteMessagePayload } from '@alkemio/matrix-adapter-lib';
 import { SendMessageToUserPayload } from '@alkemio/matrix-adapter-lib';
 import { IReaction } from '@alkemio/matrix-adapter-lib';
 import { sleep } from 'matrix-js-sdk/lib/utils';
+import { CommunicationAdminUserService } from '../communication-admin-user/communication.admin.user.service';
 
 @Injectable()
 export class CommunicationAdapter {
-  private adminUser!: IOperationalMatrixUser;
-  private matrixElevatedAgent!: MatrixAgent; // elevated as created with an admin account
-  private adminEmail!: string;
-  private adminCommunicationsID!: string;
-  private adminPassword!: string;
-  private enabled = false;
-
   constructor(
     @Inject(WINSTON_MODULE_NEST_PROVIDER)
     private readonly logger: LoggerService,
     private matrixAgentService: MatrixAgentService,
     private matrixAgentPool: MatrixAgentPool,
-    private configService: ConfigService,
     private matrixUserManagementService: MatrixUserManagementService,
     private matrixUserAdapter: MatrixUserAdapter,
-    private matrixRoomAdapter: MatrixRoomAdapter
-  ) {
-    this.adminEmail = this.configService.get(
-      ConfigurationTypes.MATRIX
-    )?.admin?.username;
-    this.adminPassword = this.configService.get(
-      ConfigurationTypes.MATRIX
-    )?.admin?.password;
-
-    this.adminCommunicationsID = this.matrixUserAdapter.convertEmailToMatrixID(
-      this.adminEmail
-    );
-
-    // need both to be true
-    this.enabled = true;
-  }
+    private matrixRoomAdapter: MatrixRoomAdapter,
+    private communicationAdminUserService: CommunicationAdminUserService
+  ) {}
 
   public async initiateMatrixConnection() {
     await this.logServerVersion();
     this.logger.verbose?.(
-      `Matrix admin identifier: ${this.adminCommunicationsID}`,
+      `Matrix admin identifier: ${this.communicationAdminUserService.adminCommunicationsID}`,
       LogContext.BOOTSTRAP
     );
-    await this.getMatrixManagementAgentElevated();
+    await this.communicationAdminUserService.getMatrixManagementAgentElevated();
   }
 
   async getCommunityRoom(
     roomID: string,
     withState = false
   ): Promise<RoomResult> {
-    // If not enabled just return an empty room
-    if (!this.enabled) {
-      const result: RoomResult = {
-        id: 'communications-not-enabled',
-        messages: [],
-        displayName: '',
-        members: [],
-      };
-      return result;
-    }
-    const matrixAgentElevated = await this.getMatrixManagementAgentElevated();
+    const matrixAgentElevated =
+      await this.communicationAdminUserService.getMatrixManagementAgentElevated();
     await this.ensureElevatedAgentIsMemberOfRoom(roomID);
 
     const matrixRoom = await this.matrixAgentService.getRoom(
@@ -321,7 +287,8 @@ export class CommunicationAdapter {
     roomID: string
   ): Promise<void> {
     try {
-      const matrixAgentElevated = await this.getMatrixManagementAgentElevated();
+      const matrixAgentElevated =
+        await this.communicationAdminUserService.getMatrixManagementAgentElevated();
       const matrixClient = matrixAgentElevated.matrixClient;
       const adminUser = this.getUserIdFromMatrixClient(matrixClient);
       await this.addUserToRoom(roomID, adminUser);
@@ -342,7 +309,8 @@ export class CommunicationAdapter {
     // when deleting a message use the global admin account
     // the possibility to use native matrix power levels is there
     // but we still don't have the infrastructure to support it
-    const matrixAgentElevated = await this.getMatrixManagementAgentElevated();
+    const matrixAgentElevated =
+      await this.communicationAdminUserService.getMatrixManagementAgentElevated();
 
     await this.ensureElevatedAgentIsMemberOfRoom(deleteMessageData.roomID);
 
@@ -383,7 +351,8 @@ export class CommunicationAdapter {
 
   async getMessageSender(roomID: string, messageID: string): Promise<string> {
     // only the admin agent has knowledge of all rooms and synchronizes the state
-    const matrixElevatedAgent = await this.getMatrixManagementAgentElevated();
+    const matrixElevatedAgent =
+      await this.communicationAdminUserService.getMatrixManagementAgentElevated();
     await this.ensureElevatedAgentIsMemberOfRoom(roomID);
     const matrixRoom = await this.matrixAgentService.getRoom(
       matrixElevatedAgent,
@@ -408,7 +377,8 @@ export class CommunicationAdapter {
 
   async getReactionSender(roomID: string, reactionID: string): Promise<string> {
     // only the admin agent has knowledge of all rooms and synchronizes the state
-    const matrixElevatedAgent = await this.getMatrixManagementAgentElevated();
+    const matrixElevatedAgent =
+      await this.communicationAdminUserService.getMatrixManagementAgentElevated();
     const matrixRoom = await this.matrixAgentService.getRoom(
       matrixElevatedAgent,
       roomID
@@ -433,56 +403,7 @@ export class CommunicationAdapter {
   }
 
   private async acquireMatrixAgent(matrixUserId: string) {
-    if (!this.enabled) {
-      throw new NotEnabledException(
-        'Communications not enabled',
-        LogContext.COMMUNICATION
-      );
-    }
     return await this.matrixAgentPool.acquire(matrixUserId);
-  }
-
-  private async getGlobalAdminUser() {
-    if (this.adminUser) {
-      return this.adminUser;
-    }
-
-    const adminExists = await this.matrixUserManagementService.isRegistered(
-      this.adminCommunicationsID
-    );
-    if (adminExists) {
-      this.logger.verbose?.(
-        `Admin user is registered: ${this.adminEmail}, logging in...`,
-        LogContext.COMMUNICATION
-      );
-      const adminUser = await this.matrixUserManagementService.login(
-        this.adminCommunicationsID,
-        this.adminPassword
-      );
-      this.adminUser = adminUser;
-      return adminUser;
-    }
-
-    this.adminUser = await this.registerNewAdminUser();
-    return this.adminUser;
-  }
-
-  public async getMatrixManagementAgentElevated() {
-    if (this.matrixElevatedAgent) {
-      return this.matrixElevatedAgent;
-    }
-
-    const adminUser = await this.getGlobalAdminUser();
-    this.matrixElevatedAgent = await this.matrixAgentService.createMatrixAgent(
-      adminUser
-    );
-
-    await this.matrixElevatedAgent.start({
-      registerTimelineMonitor: false,
-      registerRoomMonitor: false,
-    });
-
-    return this.matrixElevatedAgent;
   }
 
   private async logServerVersion() {
@@ -498,14 +419,6 @@ export class CommunicationAdapter {
         LogContext.BOOTSTRAP
       );
     }
-  }
-
-  private async registerNewAdminUser(): Promise<IOperationalMatrixUser> {
-    return await this.matrixUserManagementService.register(
-      this.adminCommunicationsID,
-      this.adminPassword,
-      true
-    );
   }
 
   async tryRegisterNewUser(email: string): Promise<string | undefined> {
@@ -533,11 +446,8 @@ export class CommunicationAdapter {
     name: string,
     metadata?: Record<string, string>
   ): Promise<string> {
-    // If not enabled just return an empty string
-    if (!this.enabled) {
-      return '';
-    }
-    const elevatedMatrixAgent = await this.getMatrixManagementAgentElevated();
+    const elevatedMatrixAgent =
+      await this.communicationAdminUserService.getMatrixManagementAgentElevated();
     const room = await this.matrixRoomAdapter.createRoom(
       elevatedMatrixAgent.matrixClient,
       {
@@ -558,10 +468,6 @@ export class CommunicationAdapter {
     roomIDs: string[],
     matrixUserID: string
   ): Promise<boolean> {
-    // If not enabled just return
-    if (!this.enabled) {
-      return false;
-    }
     try {
       await this.addUserToRooms(roomIDs, matrixUserID);
     } catch (error) {
@@ -576,10 +482,6 @@ export class CommunicationAdapter {
 
   async getCommunityRooms(matrixUserID: string): Promise<RoomResult[]> {
     const rooms: RoomResult[] = [];
-    // If not enabled just return an empty array
-    if (!this.enabled) {
-      return rooms;
-    }
 
     this.logger.verbose?.(
       `Retrieving rooms for user: ${matrixUserID}`,
@@ -589,7 +491,7 @@ export class CommunicationAdapter {
     // as protected via the graphql api
     // the possibility to use native matrix power levels is there
     // but we still don't have the infrastructure to support it
-    // const matrixAgentElevated = await this.getMatrixManagementAgentElevated();
+    // const matrixAgentElevated = await this.communicationAdminUserService.getMatrixManagementAgentElevated();
 
     // const matrixAgent = await this.matrixAgentPool.acquire(matrixUserID);
 
@@ -608,10 +510,7 @@ export class CommunicationAdapter {
 
   async getDirectRooms(matrixUserID: string): Promise<RoomDirectResult[]> {
     const rooms: RoomDirectResult[] = [];
-    // If not enabled just return an empty array
-    if (!this.enabled) {
-      return rooms;
-    }
+
     const matrixAgent = await this.matrixAgentPool.acquire(matrixUserID);
 
     const matrixDirectRooms = await this.matrixAgentService.getDirectRooms(
@@ -634,16 +533,13 @@ export class CommunicationAdapter {
     roomIDs: string[],
     matrixUserID: string
   ): Promise<boolean> {
-    // If not enabled just return
-    if (!this.enabled) {
-      return false;
-    }
     this.logger.verbose?.(
       `Removing user (${matrixUserID}) from rooms (${roomIDs})`,
       LogContext.COMMUNICATION
     );
     const userAgent = await this.matrixAgentPool.acquire(matrixUserID);
-    const matrixAgentElevated = await this.getMatrixManagementAgentElevated();
+    const matrixAgentElevated =
+      await this.communicationAdminUserService.getMatrixManagementAgentElevated();
     for (const roomID of roomIDs) {
       // added this for logging purposes
       userAgent.attachOnceConditional({
@@ -668,7 +564,8 @@ export class CommunicationAdapter {
   }
 
   async getAllRooms(): Promise<RoomResult[]> {
-    const elevatedAgent = await this.getMatrixManagementAgentElevated();
+    const elevatedAgent =
+      await this.communicationAdminUserService.getMatrixManagementAgentElevated();
     this.logger.verbose?.(
       `[Admin] Obtaining all rooms on Matrix instance using ${elevatedAgent.matrixClient.getUserId()}`,
       LogContext.COMMUNICATION
@@ -705,7 +602,8 @@ export class CommunicationAdapter {
         `[Replication] Replicating room membership from ${sourceRoomID} to ${targetRoomID}`,
         LogContext.COMMUNICATION
       );
-      const elevatedAgent = await this.getMatrixManagementAgentElevated();
+      const elevatedAgent =
+        await this.communicationAdminUserService.getMatrixManagementAgentElevated();
 
       const sourceMatrixUserIDs =
         await this.matrixRoomAdapter.getMatrixRoomMembers(
@@ -763,8 +661,11 @@ export class CommunicationAdapter {
   }
 
   public async addUserToRoom(roomID: string, matrixUserID: string) {
-    let userAgent = await this.getMatrixManagementAgentElevated();
-    if (matrixUserID !== this.adminCommunicationsID) {
+    let userAgent =
+      await this.communicationAdminUserService.getMatrixManagementAgentElevated();
+    if (
+      matrixUserID !== this.communicationAdminUserService.adminCommunicationsID
+    ) {
       userAgent = await this.matrixAgentPool.acquire(matrixUserID);
     }
 
@@ -800,9 +701,12 @@ export class CommunicationAdapter {
   }
 
   private async addUserToRooms(roomIDs: string[], matrixUserID: string) {
-    const elevatedAgent = await this.getMatrixManagementAgentElevated();
+    const elevatedAgent =
+      await this.communicationAdminUserService.getMatrixManagementAgentElevated();
     let userAgent = elevatedAgent;
-    if (matrixUserID !== this.adminCommunicationsID) {
+    if (
+      matrixUserID !== this.communicationAdminUserService.adminCommunicationsID
+    ) {
       userAgent = await this.matrixAgentPool.acquire(matrixUserID);
     }
 
@@ -877,7 +781,8 @@ export class CommunicationAdapter {
 
   async removeRoom(matrixRoomID: string) {
     try {
-      const elevatedAgent = await this.getMatrixManagementAgentElevated();
+      const elevatedAgent =
+        await this.communicationAdminUserService.getMatrixManagementAgentElevated();
       this.logger.verbose?.(
         `[Membership] Removing members from matrix room: ${matrixRoomID}`,
         LogContext.COMMUNICATION
@@ -914,7 +819,8 @@ export class CommunicationAdapter {
   async getRoomMembers(matrixRoomID: string): Promise<string[]> {
     let userIDs: string[] = [];
     try {
-      const elevatedAgent = await this.getMatrixManagementAgentElevated();
+      const elevatedAgent =
+        await this.communicationAdminUserService.getMatrixManagementAgentElevated();
       this.logger.verbose?.(
         `Getting members of matrix room: ${matrixRoomID}`,
         LogContext.COMMUNICATION
@@ -956,7 +862,8 @@ export class CommunicationAdapter {
   async updateRoomState(
     roomData: UpdateRoomStatePayload
   ): Promise<RoomDetailsResponsePayload> {
-    const elevatedAgent = await this.getMatrixManagementAgentElevated();
+    const elevatedAgent =
+      await this.communicationAdminUserService.getMatrixManagementAgentElevated();
 
     if (roomData.roomID.length === 0) {
       throw new MatrixEntityNotFoundException(
